backend/foodgram/requirements.txt
asgiref==3.5.2
Django==2.2.19
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.11
sqlparse==0.4.2
tzdata==2022.1
